remove single goal component
